pyproject.toml
[tool.poetry]
name = "swevendingmachine"
version = "0.1.0"
description = ""
authors = ["Naphong Chadha <harrajc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8"
pytest = "^7.2.1"
pyyaml = "^6.0"
sqlalchemy = "^2.0.0"
flask-sqlalchemy = "^3.0.2"
pre-commit = "^3.0.2"
pytest-cov = "^4.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
